refactor: policy 컴포넌트 페이지 리펙토링
