+++ src/components/_Home/HeroSection.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState, useRef } from 'react';
 import './css/HeroSection.css';
 import SubscriptionButton from '../common/SubscriptionButton';
-import { getBackgroundVideos } from '../utils/getBackgroundVideos';
+import { getBackgroundVideos } from '../../utils/getBackgroundVideos';
 
 const HeroSection = () => {
   const videos = getBackgroundVideos();
