+++ src/components/_Policy/PrivacyPolicySection.jsx
@@ -1,9 +1,4 @@
-import React, { useState } from 'react';
-import { IoIosArrowDown } from 'react-icons/io';
-import { IoIosArrowUp } from 'react-icons/io';
-import './css/PolicyComponent.css';
-
-const sections = [
+export const privacyPolicyData = [
   {
     title: '1. 개인정보의 수집 및 이용 목적',
     content: (
@@ -93,36 +88,3 @@ const sections = [
     ),
   },
 ];
-
-const Privacy = () => {
-  const [openSections, setOpenSections] = useState({});
-
-  const toggleSection = (index) => {
-    setOpenSections((prev) => ({ ...prev, [index]: !prev[index] }));
-  };
-
-  return (
-    <div className="privacy-container">
-      <div className="privacy-main-wrapper">
-        <h1 className="privacy-title">Haruhan 개인정보취급방침</h1>
-      </div>
-      <ul className="privacy-list">
-        {sections.map((section, index) => (
-          <li key={index} className="privacy-wrapper" onClick={() => toggleSection(index)}>
-            <div className="privacy-title-wrapper">
-              <h3>{section.title}</h3>
-              {openSections[index] ? (
-                <IoIosArrowUp width={10} height={10} />
-              ) : (
-                <IoIosArrowDown width={10} height={10} />
-              )}
-            </div>
-            {openSections[index] && <div className="privacy-content">{section.content}</div>}
-          </li>
-        ))}
-      </ul>
-    </div>
-  );
-};
-
-export default Privacy;
