+++ src/pages/Policy.jsx
@@ -1,6 +1,5 @@
-import React from 'react';
 import Layout from '../components/Layout/Layout';
-import PolicyComponent from '../components/PolicyComponent';
+import PolicyComponent from '../components/_Policy/PolicyComponent';
 
 const Policy = () => {
   return (
