+++ src/utils/getBackgroundVideos.js
@@ -1,0 +1,8 @@
+import Background1 from '../assets/videos/background1.mp4';
+import Background2 from '../assets/videos/background2.mp4';
+import Background3 from '../assets/videos/background3.mp4';
+import Background4 from '../assets/videos/background4.mp4';
+
+export const getBackgroundVideos = () => {
+  return [Background1, Background2, Background3, Background4];
+};
